backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic[email]==2.5.2
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
apscheduler==3.10.4
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
email-validator==2.1.0.post1
